+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `upgrade` command.
+
 ### Fixed
 - Minor fixes.
 - `download` command.
+- `install` command.
 
 ## [0.16.1] - 2017-08-02
 ### Fixed